feat(ts): convert features/app/ + features/ui/ to TS (PR5a)
5 file conversions. Pure SDUI: no axios, no Mongo, only lib/i18n.t()
+ moment-tz visibility filter + bus-config.data import.

- app.routes.ts: 1 sync route GET /config returning config.app.{ios,android}
- ui.buslist.ts: getBusList(lang) with discriminated-union visibility filter
- ui.campus.ts, ui.scroll.ts: static SDUI payloads with 7 lib/i18n keys
- ui.routes.ts: 3 endpoints, `req.lang as SupportedLang` cast at boundary

Type wrinkle: bus-config.data's visibility.type was inferred as wide
`string` (no `as const`), so the dateRange branch couldn't narrow
visibility.from/until from `string | undefined` to `string`. Used `!`
non-null assertions with comment justifying the runtime invariant
from bus-config.data's literal definitions — matches original .js's
direct access.

Defensive-narrowing audit: clean. Only `as SupportedLang` casts at
type-system boundaries + 2 `!` assertions with WHY comments.
